# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WatchlistCreateRequest": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "instrument_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 5000,
            "title": "Instrument Ids",
            "type": "array"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "WatchlistCreateRequest",
        "type": "object"
      },
      "WatchlistDeleteResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "deleted"
        ],
        "title": "WatchlistDeleteResponse",
        "type": "object"
      },
      "WatchlistListResponse": {
        "properties": {
          "watchlists": {
            "items": {
              "$ref": "#/components/schemas/WatchlistSummaryResponse"
            },
            "title": "Watchlists",
            "type": "array"
          }
        },
        "required": [
          "watchlists"
        ],
        "title": "WatchlistListResponse",
        "type": "object"
      },
      "WatchlistMemberResponse": {
        "properties": {
          "base_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Asset"
          },
          "company_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "integer"
          },
          "instrument_type": {
            "title": "Instrument Type",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "quote_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Asset"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "venue_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Code"
          },
          "venue_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Name"
          }
        },
        "required": [
          "instrument_id",
          "symbol",
          "instrument_type",
          "currency",
          "position"
        ],
        "title": "WatchlistMemberResponse",
        "type": "object"
      },
      "WatchlistRefreshJobResponse": {
        "properties": {
          "current": {
            "title": "Current",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "watchlist_id": {
            "title": "Watchlist Id",
            "type": "integer"
          },
          "watchlist_name": {
            "title": "Watchlist Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "watchlist_id",
          "watchlist_name",
          "status",
          "current",
          "total",
          "message",
          "started_at",
          "finished_at",
          "error"
        ],
        "title": "WatchlistRefreshJobResponse",
        "type": "object"
      },
...
      "WatchlistResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "crypto_spot_count": {
            "title": "Crypto Spot Count",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "equity_count": {
            "title": "Equity Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "instrument_types": {
            "items": {
              "type": "string"
            },
            "title": "Instrument Types",
            "type": "array"
          },
          "member_count": {
            "title": "Member Count",
            "type": "integer"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/WatchlistMemberResponse"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_refresh_supported": {
            "title": "Price Refresh Supported",
            "type": "boolean"
          },
          "reference_rate_count": {
            "title": "Reference Rate Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "member_count",
          "instrument_types",
          "equity_count",
          "crypto_spot_count",
          "reference_rate_count",
          "price_refresh_supported",
          "created_at",
          "updated_at",
          "members"
        ],
        "title": "WatchlistResponse",
        "type": "object"
      },
      "WatchlistSummaryResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "crypto_spot_count": {
            "title": "Crypto Spot Count",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "equity_count": {
            "title": "Equity Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "instrument_types": {
            "items": {
              "type": "string"
            },
            "title": "Instrument Types",
            "type": "array"
          },
          "member_count": {
            "title": "Member Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_refresh_supported": {
            "title": "Price Refresh Supported",
            "type": "boolean"
          },
          "reference_rate_count": {
            "title": "Reference Rate Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "member_count",
          "instrument_types",
          "equity_count",
          "crypto_spot_count",
          "reference_rate_count",
          "price_refresh_supported",
          "created_at",
          "updated_at"
        ],
        "title": "WatchlistSummaryResponse",
        "type": "object"
      },
      "WatchlistUpdateRequest": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "instrument_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 5000,
            "title": "Instrument Ids",
            "type": "array"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "WatchlistUpdateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backtesting and factor analysis API",
    "title": "Trading Statistics Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/watchlists": {
      "get": {
        "operationId": "listWatchlists",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Watchlists",
        "tags": [
          "watchlists"
        ]
      },
      "post": {
        "operationId": "createWatchlist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Watchlist",
        "tags": [
          "watchlists"
        ]
      }
    },
...
    "/watchlists/refresh-jobs/{job_id}": {
      "get": {
        "operationId": "getWatchlistRefreshJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistRefreshJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Watchlist Refresh Job",
        "tags": [
          "watchlists"
        ]
      }
    },
    "/watchlists/{watchlist_id}": {
      "delete": {
        "operationId": "deleteWatchlist",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Watchlist",
        "tags": [
          "watchlists"
        ]
      },
      "get": {
        "operationId": "getWatchlist",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Watchlist",
        "tags": [
          "watchlists"
        ]
      },
      "put": {
        "operationId": "updateWatchlist",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Watchlist",
        "tags": [
          "watchlists"
        ]
      }
    },
    "/watchlists/{watchlist_id}/refresh": {
      "post": {
        "operationId": "refreshWatchlistPrices",
        "parameters": [
          {
            "in": "path",
            "name": "watchlist_id",
            "required": true,
            "schema": {
              "title": "Watchlist Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistRefreshJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Watchlist Prices",
        "tags": [
          "watchlists"
        ]
      }
    }
  }
}